requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msql==1.2.2
mypy==0.812
mypy-extensions==0.4.3
mysql-connector==2.2.9
PyMySQL==1.0.2
SQLAlchemy==1.4.6
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
